Update README.md
This is a function and game called GuessTheNumber()The function will generate a random number between 0 and 10 and the user will be prompted to guess the number. If the guess is too high the function will respond "Too High" if it is too low, it will respond "Too Low" and if the user guesses correctly it will say "You Got It!" IF the user does not get it, they will be prompted to guess again until they get it correct. the Lower bound is 0, the upper bound is 10, and the seed is NULL. This function also has a series of outputs RandomNumber; the random number generated, NumGuesses; the number of guesses the user took, and Guesses; which is a vector of all the guesses.